control-api/requirements.txt
fastapi
uvicorn
aiosqlite
paho-mqtt
pydantic
